1. fix issue #1 : error of compiling phalcon 2. using trusty64 as base VM os. 3. add install-mariadb scripts from *Homestead* project 4. clean up some code

+++ scripts/install-base-env.sh
@@ -1,32 +1,41 @@
 #!/usr/bin/env bash
 
+# Check If base env Has Been Installed
+
+if [ -f /home/vagrant/.env ]
+then
+    echo "Base Env already installed."
+    exit 0
+fi
+
+touch /home/vagrant/.env
 # Pre-requisites
-sudo apt-get -y update
-sudo apt-get --no-install-recommends -y install python-software-properties openssl curl git-core sqlite3 pkg-config build-essential
+apt-get -y update
+apt-get --no-install-recommends -y install python-software-properties openssl curl git-core sqlite3 pkg-config build-essential
 
 
 # libs
-sudo apt-get -y install libevent-dev
+apt-get -y install libevent-dev
 
 
 # servers
-sudo apt-get -y install nginx-light rsync supervisor
+apt-get -y install nginx-light rsync supervisor
 cp -f /vagrant/files/etc/nginx/sites-available/default /etc/nginx/sites-available/
 /etc/init.d/nginx restart
 
 
 
 # php
-sudo apt-get -y install php-apc php-pear php5-cgi php5-cli php5-common php5-curl php5-dev php5-fpm php5-gd php5-mcrypt php5-sqlite php5-pgsql
+apt-get -y install php-apc php-pear php5-cgi php5-cli php5-common php5-curl php5-dev php5-fpm php5-gd php5-mcrypt php5-sqlite php5-pgsql
 
 # install fpm's php.ini
-sudo cp -f /vagrant/files/etc/php5/fpm/php.ini /etc/php5/fpm/
+cp -f /vagrant/files/etc/php5/fpm/php.ini /etc/php5/fpm/
 
 # php build-dep
-sudo apt-get -y build-dep php5
+apt-get -y build-dep php5
 
 # tools
-sudo apt-get -y install zip vim re2c
+apt-get -y install zip vim re2c
 
 #
 # composer
@@ -38,5 +47,5 @@ chown -R vagrant:vagrant /home/vagrant/bin
 
 # copy files to home
 cp -f /vagrant/files/dot/.bash_aliases /home/vagrant/
-sudo chown vagrant:vagrant .bash_aliases
+chown vagrant:vagrant .bash_aliases
